Added animations to the player
edited the original sprite sheet to be able to use a colour key with non conflicting colours that are present on the sprite and extracted each part of every animation from the sprite sheet. created an animate function which depending on the status of the player changes the animation that is being looped on screen. This was done through the input function by adding a status to each key press