Added the start of interaction with the map
If the player now swings their axe on a tree with a plum the plum gets removed. In player.py I updated the use tool method to call the break tree function in sprites.py which lowers the trees health and removes a single plum if present. A new sprite group was created in world.py for the ability to detect when a player is against a tree for knowing whether a hit against a tree with the axe registers or not. timer.py was updated as whenever a plum crossed with the players hitbox before the plum would be removed and so by adding the "and" statement to the if statement and placing the self deactivate after the conditional this problem was fixed