Updated sprites.py to now display a small animation on the tree when being broken
The tree now flashes white when being broken. This was done by calling the particle class in the CheckBreak method. I needed to add a new parameter to the particles class so that the item could be checked whether a plum was being broken or the tree and use the correct height and width amounts for each situation.